requirements.txt
blinker==1.7.0
cachelib==0.10.2
click==8.1.7
Flask==3.0.1
Flask-Session==0.6.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
python-dateutil==2.8.2
PyYAML==6.0.1
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
